Aded 2 Layer Delete Alert Notification In ManageUsers Page

+++ src/pages/Dashboard/AdminDashboard/ManageUsers/ManageUsers.jsx
@@ -4,6 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import { MdAddModerator, MdAdminPanelSettings } from "react-icons/md";
 import { PiStudentFill } from "react-icons/pi";
 import { FaRegTrashCan } from 'react-icons/fa6';
+import Swal from 'sweetalert2';
 
 const ManageUsers = () => {
 
@@ -30,11 +31,30 @@ const ManageUsers = () => {
     }
 
     const deleteUser = user => {
-        axios.delete(`/users/${user._id}`)
-            .then(res => {
-                refetch();
-                console.log('after delete user', res.data);
-            })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                axios.delete(`/users/${user._id}`)
+                    .then(res => {
+                        if (res.data.deletedCount) {
+                            refetch();
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your parcel request has been deleted.",
+                                icon: "success"
+                            });
+                        }
+                    })
+            }
+        });
     }
 
     return (
